+++ provider/dataset/in_memory_dataset.go
@@ -1,3 +1,10 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright 2024 FeatureForm Inc.
+//
+
 package dataset
 
 import (
@@ -23,7 +30,7 @@ func (ds *InMemoryDataset) Location() pl.Location {
 }
 
 func (ds *InMemoryDataset) Iterator(ctx context.Context) (Iterator, error) {
-	return &InMemoryIterator{data: ds.data, index: -1}, nil
+	return &InMemoryIterator{data: ds.data, schema: ds.schema, index: -1}, nil
 }
 
 func (ds *InMemoryDataset) Schema() (types.Schema, error) {
